LPD-27759 Use the same approach used in PollsPortletIdToDDMPortletIdUpgradeProcess, it means that if an upgrade process test is almost the same as an already existing one it should extend it instead of creating a base class
