+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v4_3_3/test/WizardDDMStructureLayoutCTUpgradeProcessTest.java
@@ -6,9 +6,39 @@
 package com.liferay.dynamic.data.mapping.upgrade.v4_3_3.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.change.tracking.test.util.BaseCTUpgradeProcessTestCase;
+import com.liferay.dynamic.data.mapping.constants.DDMStructureConstants;
+import com.liferay.dynamic.data.mapping.model.DDMForm;
+import com.liferay.dynamic.data.mapping.model.DDMFormInstance;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
-import com.liferay.dynamic.data.mapping.upgrade.BaseDDMStructureLayoutCTUpgradeProcessTestCase;
+import com.liferay.dynamic.data.mapping.model.DDMStructure;
+import com.liferay.dynamic.data.mapping.model.DDMStructureLayout;
+import com.liferay.dynamic.data.mapping.service.DDMStructureLayoutLocalService;
+import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
+import com.liferay.dynamic.data.mapping.storage.StorageType;
+import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
+import com.liferay.dynamic.data.mapping.util.DDMUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.change.tracking.CTModel;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.change.tracking.CTService;
+import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
+import com.liferay.portal.kernel.test.util.GroupTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
+import com.liferay.portal.kernel.test.util.TestPropsValues;
+import com.liferay.portal.kernel.upgrade.UpgradeProcess;
+import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portal.test.rule.Inject;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
+import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
+import com.liferay.portal.upgrade.test.util.UpgradeTestUtil;
 
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.runner.RunWith;
 
 /**
@@ -16,16 +46,94 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 public class WizardDDMStructureLayoutCTUpgradeProcessTest
-	extends BaseDDMStructureLayoutCTUpgradeProcessTestCase {
+	extends BaseCTUpgradeProcessTestCase {
+
+	@ClassRule
+	@Rule
+	public static final AggregateTestRule aggregateTestRule =
+		new AggregateTestRule(
+			new LiferayIntegrationTestRule(),
+			PermissionCheckerMethodTestRule.INSTANCE);
+
+	@Before
+	public void setUp() throws Exception {
+		_group = GroupTestUtil.addGroup();
+
+		_serviceContext = ServiceContextTestUtil.getServiceContext(
+			_group.getGroupId(), TestPropsValues.getUserId());
+	}
 
 	@Override
+	protected CTModel<?> addCTModel() throws Exception {
+		DDMForm ddmForm = DDMFormTestUtil.createDDMForm(
+			RandomTestUtil.randomString());
+
+		DDMStructure ddmStructure = _ddmStructureLocalService.addStructure(
+			TestPropsValues.getUserId(), _group.getGroupId(), 0,
+			PortalUtil.getClassNameId(DDMFormInstance.class.getName()), null,
+			RandomTestUtil.randomLocaleStringMap(), null, ddmForm,
+			_createDDMFormLayout(ddmForm), StorageType.DEFAULT.toString(),
+			DDMStructureConstants.TYPE_DEFAULT, _serviceContext);
+
+		return ddmStructure.fetchDDMStructureLayout();
+	}
+
+	@Override
+	protected CTService<?> getCTService() {
+		return _ddmStructureLayoutLocalService;
+	}
+
 	protected String getPaginationMode() {
 		return DDMFormLayout.WIZARD_MODE;
 	}
 
-	@Override
 	protected String getSchemaVersion() {
 		return "v4_3_3";
 	}
+
+	@Override
+	protected void runUpgrade() throws Exception {
+		UpgradeProcess upgradeProcess = UpgradeTestUtil.getUpgradeStep(
+			_upgradeStepRegistrator,
+			"com.liferay.dynamic.data.mapping.internal.upgrade." +
+				getSchemaVersion() + ".DDMStructureLayoutUpgradeProcess");
+
+		upgradeProcess.upgrade();
+	}
+
+	@Override
+	protected CTModel<?> updateCTModel(CTModel<?> ctModel) throws Exception {
+		DDMStructureLayout ddmStructureLayout = (DDMStructureLayout)ctModel;
+
+		return _ddmStructureLayoutLocalService.updateStructureLayout(
+			ddmStructureLayout.getStructureLayoutId(),
+			_createDDMFormLayout(
+				DDMFormTestUtil.createDDMForm(RandomTestUtil.randomString())),
+			_serviceContext);
+	}
+
+	private DDMFormLayout _createDDMFormLayout(DDMForm ddmForm) {
+		DDMFormLayout ddmFormLayout = DDMUtil.getDefaultDDMFormLayout(ddmForm);
+
+		ddmFormLayout.setPaginationMode(getPaginationMode());
+
+		return ddmFormLayout;
+	}
+
+	@Inject(
+		filter = "(&(component.name=com.liferay.dynamic.data.mapping.internal.upgrade.registry.DDMServiceUpgradeStepRegistrator))"
+	)
+	private static UpgradeStepRegistrator _upgradeStepRegistrator;
+
+	@Inject
+	private DDMStructureLayoutLocalService _ddmStructureLayoutLocalService;
+
+	@Inject
+	private DDMStructureLocalService _ddmStructureLocalService;
+
+	@DeleteAfterTestRun
+	private Group _group;
+
+	private ServiceContext _serviceContext;
 
 }
